backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
openai>=1.0.0
huggingface-hub>=0.31.0
requests>=2.31.0
pandas==2.2.3
openpyxl==3.1.5
pdfplumber==0.11.5
chromadb>=0.5.0
sentence-transformers>=3.0.0
langchain-text-splitters>=0.3.0
python-docx==1.1.2
python-multipart>=0.0.6
sympy==1.13.3
matplotlib==3.10.0
scikit-learn==1.6.1
joblib==1.4.2
scipy==1.15.1
numpy==2.2.1
firebase-admin>=6.2.0
redis[hiredis]>=5.0.0
PyYAML>=6.0.0
mypy>=1.20.0
pytest>=9.0.0
pytest-asyncio>=0.23.0
google-api-python-client>=2.0.0
pypdf>=4.0.0
